Move supporting code into module
